Fix ed commands in makehelp.sh
If an 's' command is part of a 'g' command list, the replacement of the
's' command cannot use an escaped newline to split the line.  The three
's' commands that split a line are no longer part of a 'g' command and
are applied to all lines by incorporating the match expression of their
respective 'g' command.